refactor to a single settings page
